重构SeverTool
