+++ src/LazyAPI/PluginContainer.cs
@@ -10,7 +10,7 @@ public class PluginContainer : LazyPlugin
 {
     public override string Author => "cc004 & members of UnrealMultiple";
 
-    public override Version Version => new Version(1, 0, 1, 0);
+    public override Version Version => new Version(1, 0, 2, 0);
 
     public PluginContainer(Main game) : base(game) { }
     public override void Initialize()
