+++ src/LazyAPI/README.md
@@ -6,9 +6,113 @@
 
 ## 命令行参数
 - `-culture zh|en` 此命令行参数用于切换生成配置文件的语言，默认跟随ts语言。
+
+## 自动注册指令详解
+
+*类声明*
+如此声明一个类，给类添加一个CommandAttribute，LazyAPI会在插件实例化阶段进行检测并进行下一步操作。
+```csharp
+[Command("test")]
+public class TestCommand
+{
+
+}
+```
+
+*指令方法*
+声明的方法必须为静态且公开的，第一个参数必须是`CommandArgs`，否则不会加载!
+```csharp
+[Command("test")]
+public class TestCommand
+{
+	public static void Run(CommandArgs args)
+	{
+		...方法内部实现...
+	}
+}
+```
+
+*子命令*
+下方示例被加载后`test`命令会出现两个子命令`/test run` 和 `/test go`，默认情况下子命令以小写方法名命名，第二条则是因为添加了`Alias`Attribute。
+```csharp
+[Command("test")]
+public class TestCommand
+{
+	public static void Run(CommandArgs args)
+	{
+		...方法内部实现...
+	}
+
+	[Alias("go")]
+	public static void Execute(CommandArgs args)
+	{
+		...方法内部实现...
+	}
+}
+```
+
+如果不希望成为子命令则只需要添加一个`MainAttribute`的Attribute，这样当你运行`/test`就不存在子命令可以直接运行
+```csharp
+[Command("test")]
+public class TestCommand
+{
+	[Main]
+	public static void Execute(CommandArgs args)
+	{
+		...方法内部实现...
+	}
+}
+```
+
+*其他Attribute*
+`PermissionsAttribute`为指令附加了权限，如此一来便不需要在使用HasPermission(xx)来进行判断
+`KindAttribute`的存在则是为了处理翻页Page的情况，后续将会提到。
+```csharp
+[Command("test")]
+[Permissions("test")]
+public class TestCommand
+{
+	[Kind]
+	[RealPlayer]
+	[Permissions("test.use")]
+	public static void Execute(CommandArgs args)
+	{
+		...方法内部实现...
+	}
+}
+```
+*参数匹配*
+LazyAPI的指令系统拥有参数匹配功能
+以下示例中，方法多了一个`string`和`int`类型参数，LazyAPI在处理指令时，会从args.Parameters中找到这些参数并尝试转换。
+并且在默认情况下，参数是严格匹配的，列如输入指令的参数和声明函数参数数量不同时不会执行，若参数转换失败同样如此。
+可自动处理的参数类型:`int`，`long`，`TSPlayer`，`bool`，`string`，`UserAccount`，`DateTime`
+```csharp
+[Command("test")]
+[Permissions("test")]
+public class TestCommand
+{
+	[Permissions("test.use")]
+	public static void Execute(CommandArgs args, string name, int age)
+	{
+		...方法内部实现...
+	}
+}
+```
+*KindAttribute*
+此标签是为了处理那些特殊指令，例如:`test list [页码]`，因为参数匹配限制，所以添加了这个Attribute，你只需给方法添加它，那么LazyAPI就不会那么严格要求参数数量，只要大于方法参数数量即可。
+
+## Rest
+关于Rest LazyAPI的处理与指令基本相同，可以参考ServerTool中的示例。
+
 ## 更新日志
 
 ```
+v1.0.2.0
+1.修复在构造函数阶段使用TShock.Log
+2.添加FindCommand
+3.添加RestPathAttribute，FindAttribute
+4.RestHelper可以自定义method路径
+
 v1.0.1.0
 修复`TShockAPI.I18n.TranslationCultureInfo`为空时, 无法正确生成配置
